remove v1 migration and keep id as a serial type(not bigserial)

+++ src/main/java/machinosoccerweb/registration/repositories/EmailConfRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EmailConfRepository extends JpaRepository<EmailConf, Long> {
+public interface EmailConfRepository extends JpaRepository<EmailConf, Integer> {
 }
